+++ rio_tiler_pds/modis/__init__.py
@@ -1,1 +1,3 @@
 """rio-tiler-pds.modis"""
+
+from .utils import sceneid_parser  # noqa