+++ rio_tiler_pds/cbers/__init__.py
@@ -1,1 +1,3 @@
 """rio-tiler-pds.cbers"""
+
+from .utils import sceneid_parser  # noqa