+++ rio_tiler_pds/landsat/__init__.py
@@ -1,1 +1,3 @@
 """rio-tiler-pds.landsat"""
+
+from .utils import sceneid_parser  # noqa