+++ rio_tiler_pds/landsat/__init__.py
@@ -1,3 +1,4 @@
 """rio-tiler-pds.landsat"""
 
-from .utils import sceneid_parser  # noqa
+from rio_tiler_pds.landsat import aws  # noqa
+from rio_tiler_pds.landsat.utils import sceneid_parser  # noqa