+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Next (TBD)
+
+* add sentinel-2 product ID parsing.
+
 ## 0.4.1 (2020-11-24)
 
 * update for rio-tiler 2.0.0rc3